detect-ssl-version: move unittests to tests/
